Remove Validator type to simplify ValidatorSet
Previously, ValidatorSet kept a map of NodeId to Validator but the
Validator type only has the NodeId and Stake. We should just map the
NodeId to Stake in ValidatorSet and remove the Validator type.

If in the future, there is more information to track about a Validator,
I think we should add a ValidatorProperty type that has Stake and
whatever future information gets added, and the mapping stays the same
in ValidatorSet.