Change the id of the NopSignature
Previously, the id of the NopSignature was using a random number
generated at runtime. This is not ideal, as it makes the MonadEvent
impossible to compare when we replay the event log during the
visualization. Thus, in order to provide a consistent way of doing
equality comparisons, we now use the hash of the message as the id
in the signature function.