Fix `TestSigner`
Using the `TestSigner` is more complex than it needs to be. The test code
needs to hash the object it wants to sign first, then pass the msg hash
to the `TestSigner`.

Instead of passing in the object and msg hash, the `sign_object` method
accepts a `Hasher` generic and hash the object in the function, and then
create the signature on the message hash.

All the use cases right now are create valid signatures. If one wants to
create an invalid signature, it can sign the object with a different keypair.
It produces the same verification error (`InvalidAuthor`) as if the msg
hash is wrong.