Add StateRoot hashes to consensus and seq_num to blocks
In order for proposals to be validated, the state root hash of a
previously executed block needs to be checked. The root hash to be
checked is a fixed delta relative to the current block being proposed.

There is a difference between the number of blocks executed and the
round number in blocks. Because of timeouts, we can be on round 1000
while the block in the proposal is only the 3rd proposed block.
The state-root is relative to executed blocks, not rounds, so a sequence
number is introduced to identify which state-root hash to check.

The state-root to check could also be determined via the ordering of
state root hashes (queue) but sequence id helps solve the problem of
identifying which hash to use if a proposal from a high round arrives
which does not extend a known branch of the blocktree.
If the round is much higher but the sequence number is still within a
range of available state root hashes, the node can still validate the
proposal, accept the proposal and blocksync the missing blocks. Without
the sequence number, its unknown how many blocks are missing and whether
this node has the state roots necessary to verify those blocks and
therefore whether blocksync is sufficient to catch up or if a more
involved state sync is necessary.