Make monad-testutil functions generic
Previously, monad-testutil had type aliases defined at the top of the
file, which the functions referenced directly. This commit changes all
of those functions to be properly generic.

One consequence of this is that the generic type specification bleeds
up into all of the tests that depend on moand-testutil. This was a
conscious decision - because the alternative would be to define
additional helper functions in monad-testutil with default values for
some of the generic parameters. I chose not to take that approach
because as we start to test more and more permutations of different
executors/validators/signatures/etc, the usefulness of preset common
defaults will start to degrade.