`SignatureCollection` serialization
The top-level state requires the events to be serializable for the persistence
logger. MonadEvent is generic over `MessageSignature` and `SignatureCollection`,
so we need to add `(de)serialize` to `SignatureCollection` similar to how
it's done for MessageSignature.

There are a few bit that are inelegant. `BitVec`'s serialization is more
complex than copying the underlying container (Vec<u8>). It contains other
information such as bit length, ordering, etc. So I used its `serde::Serialize`
implementation to serialize the bit vector.

The other bit is that `SignatureCollection::(de)serialize` function is pulling
in the protobuf prototypes, because the SignatureCollection has more than
1 field. So I had to resolve to prototypes to serialize them.