Add config param for proposal behaviour when missing blocks
In certain situations where we are missing blocks in the pending
blocktree (don't have a chain of blocks back to the root), we still
might have enough information to try and propose a block. The block may
have a higher chance to fail, but proposing in that situation gives the
validator a chance to try and push a block through.