Switch `TimeoutCertificate` to `SignatureCollection`
`TimeoutCertificate` was created with a vector of signatures over the current
round and the high qc round. The size of the certificate is thus proportional
to the number of signers on it.

We note that nodes are likely to be locked on a small number of high qcs,
so we can leverage aggregate signatures to make TC more compact and
potentially faster to verify.

`TimeoutMessage` now carries a signature created with the certificate keypair.
We bucket the timeout signatures by `high_qc_round` and aggregate signatures
in the same bucket when constructing the TC.