Allow MockMempool to initialize with a seed
MockMempool currently return the same bytes as long as the original
fetch_txs requested non-empty transactions.

For TwinBFT's testing, is very useful to have mock_mempool to return
different txs (or in this setting, simply bytes) given the same round
condition thus generating a proposal that is similar but not identical.

Thus, a seed is added to the initialization process of MockMempool,
the default value when using MockMempoolConfig::default() is set to 1