Abstract swarm relationship into SwarmRelation
This pr abstract majority of "where" clauses and Generic to a single
trait, thus allow easier modification in future and readability.

With the change of converting State in Nodes to MonadState, the amount
of generic and the relationship between generic is pilling up, This pr
would greatly reduce the difficulty for such change and also any future
changes.