Improve monad-testground qdisc config
This addresses two issues in our tc config generator. Firstly, the drr
qdisc does not work properly as a parent of a netem qdisc. This is
because drr only polls from an active list of its child classes that it
maintains. Once a child netem qdisc is marked as inactive, it only gets
marked as active once another skb is enqueued onto it.

Secondly, ingress traffic rate limiting was broken. This is resolved by
using an IFB (Intermediate Functional Block) device, which allows for
classful qdiscs to be used on the ingress path.