Add basic evidence storage interface
Evidence is added to monad-types. Collected evidence can be converted to
a cmd and sent to evidence executor for long-term storage

trait State now defines a new type called Violation, which is the core of
Evidence being sent around. Example enum (which may be the only enum) is
ConsensusViolation

Thus, the format is: Evidence(node_id, Violation[proof_a, proof_b, ...])

ConsensusViolation stores the message signature as bytes. Internally,
individual nodes never re-verify signature but when sharing/sending
evidence with other nodes, signature serves as a proof of violation

At the moment, only NopEvidenceStorage exists as a demo executor

ProofOfVote is used in the next commit as an example