Move BlockSyncMsgs out of ConsensusMsg
Block sync is not core consensus message, so moving it up 1 layer into
MonadMessage.

VerifiedMessage is converted from a transparent single field struct to a
enum, and thus we lose the `RefCast` on it.

The Identifiable trait used the consensus signature as the Id. I'm
proposing to remove it as it's unused after we removed gossip unpublish.
We need an alternative implementation (maybe msg hash) if we want to
keep it.