Add count getter to CounterLayer
Our tests used `traced_test` with `logs_assert` on counter logs to
detect if certain events occured. When the test fails, it emits a
massive log, taking forever to print and we can't filter it by log level
because counter logs are on the trace log level.

This PR allows us to access the counter directly, making writing
assertions like these easier. They run faster than logs_assert as well.

Tracing subscriber cannot be unset once set, so we can't have multiple
tests in the same file all use this functionality. This can be fixed if
we create a top-level scope/span for each test case, and make counter
aware of that. That's deferred for now.