Version check the consensus protocol messages
The consensus protocol and algorithm could change without any of the
protocol messages acutally changing their structure so we should have a
way of checking for versions. This aids in upgradability of the
consensus algorithm.

The consensus protocol is neither backwards nor forwards compatible so
we just use an exact string match to determine if a message should be
validated.

This patch also includes the structure for a client major-minor version
number but it is not checked yet.