Write blockdb from blockwatch example
Finalized blocks and some other information need to be retrieved by the
rpc server so a database for blocks and transaction hashes is useful.
This patch provides an example of the blockdb storage via lmdb,
implemented in the blockwatch tool.

This is not the permanent solution for populating blockdb; the execution
ledger executor should ultimately be responsible for doing this.