Add TestMode to monad-node
monad-node by default runs in production mode. The test mode adds
commands to inject different behaviors/updaters, at the cost of some
updaters getting dynamically dispatched.

I switched to this approach cause it's too hard to keep two binaries in
sync for the most part, but different in subtle ways. That might be
easier to do after we have fewer commits to the binary.