Fix lmdb database open usage
lmdb database opens (mdb_dbi_open) only need to be created once per
environment as the handle is then shared within the environment until
closed. The reference to the handle is just a u32 so multiple cursors
in different threads can access it and let the underlying lmdb
environment handle locking.

mdb_dbi_open should not be called from multiple concurrent transactions
in the same process