Try commit blocks when path_to_root is repaired
Previous implementation only checks commit condition in handle proposal.
When a block sync message or out-of-order proposal is received, we add
the block to block tree, but do not attempt to commit.

This leads to issues with consensus restarts near epoch boundaries.
Consider a scenario where, by the time the restarting node finishes
syncing via block sync (requested the boundary block but not committing
it), the network has moved on to the next epoch. The node never gets the
chance to commit the boundary block and is stuck. This PR addresses that