NullBlock/LivenessBlock seqnum change
A node proposes an empty block if don't have the execution state root,
delay blocks ago. Empty blocks now use the same sequence number and
state root as their parent. They are committed to consensus ledger, but
not to eth ledger. This provides a nice property that if we observe
QC(N), block N-delay is guaranteed to be committed on a supermajority of
nodes. We can then state sync to N-delay.

Some implications of the change

- Epoch is now measured in number of non-empty blocks
- We wouldn't issue reward for empty blocks, as they are not committed
to execution. Otherwise nodes can always produce empty blocks, get
rewards, and never move in to a new epoch