Remove WAL replay
We previously were using the WAL as part of our canonical node recovery
mechanism. We now only use the WAL as a debug tool. Because of this, we
no longer need the special-cased Executor::replay functions. These
functions were previously used to try and make sure that the hidden
executor states were seeded properly upon replay, so that we wouldn't
get events double-emitted.