Switch monad-raptorcast over to custom Raptor encoder
This PR imports an in-house developed Raptor code encoder and switches
monad-raptorcast over to use this encoder.

A corresponding Raptor decoder has also been developed, and a subsequent
PR will switch monad-raptorcast over to use that decoder as well.

We developed this custom Raptor encoder/decoder pair because of concerns
about the performance of the available open-source Raptor decoder
implementations, and because we are using a non-standard, non-systematic
variant of an existing standard Raptor code (being the R10 code described
in RFC 5053), for which there aren't any readily available open-source
implementations.

This PR causes no test failures for any of the pre-existing tests in the
monad-bft repository, however, we don't have any tests that verify that
when monad-raptorcast communicates with itself the originally submitted
application message comes out unscathed at the other end, so this doesn't
prove that the Raptor encoder in the PR is compatible with what we are
currently using -- and indeed, if I modify the encoder to produce garbage
output, the relevant tests still pass.  However, since we're not using
monad-raptorcast yet, this is likely OK, and we don't need to expend
resources on ensuring compatibility at this time.

All code in `src/r10/*.rs` should comply with RFC 5053, and the code for
our non-systematic variant of RFC 5053 lives in `src/r10/nonsystematic/`.
We stick as closely as possible to what the systematic code variant does
in the areas where we don't necessarily need to make any changes, and we
also retain the use of the systematic indices defined in section 5.7, not
for the reason that we need the A matrix to be invertible so that the
intermediate symbol generation step works, as we don't generate the
intermediate symbols in the same way in the non-systematic code variant,
but for the reason that that makes the first N consecutive encoded symbols
guaranteed to be able to decode the encoded message, which helps when
stress-testing the decoder to make sure that it is actually able to decode
all encoded symbol subsets that are theoretically decodable, and it allows
setting the redundancy level to 0 for testing purposes.

RFC 5053 doesn't specify any test vectors, and we have not tested (the
systematic variant of) this encoder/decoder against other implementations
of this Raptor code, so we can't verify directly that our implementation
of the RFC 5053 parameter generation etc. is 100% correct, but the A
matrices generated by this encoder are invertible for all legal values
of K (= the number of source symbols), as they should be according to the
RFC, which is a good indicator that we've implemented the relevant parts
of the standard correctly -- and that is important because we want the
favorable overhead-failure properties of the RFC 5053 code to apply to
our code.

Even though we don't know for sure that the new Raptor encoder produces
exactly the same output as the existing one, it's at least a bit faster.
The existing one (git commit 557cc74987cd) gets these numbers on my laptop:

```
$ cargo bench -p monad-raptorcast Encoding -- --nocapture
[...]
encoder/decoder/Encoding
                        time:   [24.353 ms 24.500 ms 24.698 ms]
                        thrpt:  [154.45 MiB/s 155.70 MiB/s 156.64 MiB/s]
```

And the encoder in this PR:

```
encoder/decoder/Encoding
                        time:   [22.146 ms 22.181 ms 22.227 ms]
                        thrpt:  [171.63 MiB/s 171.98 MiB/s 172.25 MiB/s]
                 change:
                        time:   [-9.7825% -9.6215% -9.4243%] (p = 0.00 < 0.05)
                        thrpt:  [+10.405% +10.646% +10.843%]
                        Performance has improved.
```

The encoder uses a const generic for the symbol (chunk) length, which
allows the compiler to unroll and cleanly auto-vectorize all the symbol
XORing that we need to do, but which is a little inconvenient for
monad-raptorcast, as it restricts us to a single Merkle tree depth.  For
now we always set the Merkle tree depth to 6, and I have some ideas for
how to handle this better in the future.

(Note that for the current monad-dataplane GSO size and Raptorcast
parameters, the symbol length comes out to be 1257 bytes, and we might
want to reduce this by a few bytes so that it is a multiple of 8 bytes,
or even of 16 bytes.)