monad-dataplane: Restart system calls on EINTR
When attaching to the monad-raptorcast `service` example using
`strace -p <pid>`, the example dies with a panic that looks like this:

    thread 'thread '<unnamed><unnamed>' panicked at ' panicked at monad-dataplane/src/event_loop.rsmonad-dataplane/src/event_loop.rs:162::16255:
    :called `Result::unwrap()` on an `Err` value: Os { code: 4, kind: Interrupted, message: "Interrupted system call" }55
    :
    note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
    called `Result::unwrap()` on an `Err` value: Os { code: 4, kind: Interrupted, message: "Interrupted system call" }
    Aborted (core dumped)

This happens because monad-raptorcast dies when a system call returns
EINTR instead of restarting the system call, and this makes debugging
problems using strace somewhat difficult.

This commit adds a helper function `retry_eintr()` which wraps the
invocation of a system call and which automatically restarts that system
call if EINTR is returned.  With this change, the `service` example
successfully survives being attached to using `strace -p <pid>` several
times in a row.