monad-raptorcast: Reorder post-proof header fields to improve alignment
Without the changes in this PR, the part of the RaptorCast header
after the Merkle proof section looks like this:

             0        4        8        c
    0x0000:  xx                                   chunk_merkle_leaf_idx
               xxxxxx xxxxxxxx xxxxxxxx xxxxxxxx  chunk_recipient
    0x0010:  xxxxxxxx xx
                        xxxx                      chunk_id
                            xx xx                 chunk_len
                                 xxxxxx xx......  data

Note the lack of proper alignment of several of the fields, and how the
data field has a length that is congruent to 3 mod 4.

This PR moves the chunk_merkle_leaf_idx byte to be after the
chunk_recipient field, and removes the chunk_len field (since we can
infer the chunk length from the length of the packet) and replaces it
with a single padding byte.  This makes the layout look like this:

             0        4        8        c
    0x0000:  xxxxxxxx xxxxxxxx xxxxxxxx xxxxxxxx  chunk_recipient
    0x0010:  xxxxxxxx
                      xx                          chunk_merkle_leaf_idx
                        xx                        reserved
                          xxxx                    chunk_id
                               xxxxxxxx ........  data

This looks a bit better, and this way, all fields are naturally aligned,
and the data field is a multiple of 4 bytes long.