monad-raptorcast: Gracefully handle Raptor encoder initialization errors
When there is an error initializing a Raptor encoder context, the
current raptorcast code panics, as it calls unwrap() on the Result
that is returned by monad_raptor::Encoder::new().

Initializing a Raptor encoder can fail if the application message to
be encoded is too small (corresponding to fewer than the R10 minimum
number of 4 encoded symbols) or too large (more than the R10 maximum
number of 8192 encoded symbols).  This is unlikely to happen, but if
it does, we probably don't want to crash the process.

This commit adds a test which without the included fix panics with:

---- encoder_error stdout ----
thread 'encoder_error' panicked at /home/buytenh/git/monad-bft/monad-raptorcast/src/udp.rs:404:86:
called `Result::unwrap()` on an `Err` value: Custom { kind: InvalidInput, error: "number of source symbols 9750 not in range 4..=8192" }
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

And with the included fix, it logs a message but does not panic:

$ RUST_LOG=warn cargo test -p monad-raptorcast
[...]
running 1 test
2024-09-09T20:29:18.066602Z  WARN monad_raptorcast::udp: unable to create Encoder, dropping message err=Custom { kind: InvalidInput, error: "number of source symbols 9750 not in range 4..=8192" }
test encoder_error ... ok