Gracefully handle Raptor decoder initialization errors in raptorcast
Before this commit, it was possible to panic a RaptorCast instance by
sending it an encoded symbol that is properly signed and formatted but
which indicates an app_message_len that is oversized, i.e. corresponds to
more than the maximum number of R10 source symbols (8192) worth of data.

The test added by this commit panics the RaptorCast receive path if the
fix added by this commit is backed out, as follows:

    $ cargo test -p monad-raptorcast oversized -- --nocapture
    [...]
    running 1 test
    thread 'tokio-runtime-worker' panicked at /home/buytenh/git/monad-bft/monad-raptorcast/src/udp.rs:158:69:
    called `Result::unwrap()` on an `Err` value: Custom { kind: InvalidInput, error: "number of source symbols 9741 not in range 4..=8192" }
    note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
    error: test failed, to rerun pass `-p monad-raptorcast --test malformed_inputs`

With the fix added by this commit included, we no longer panic, and
instead log a message that looks like this:

    2024-09-19T15:04:29.332538Z  WARN monad_raptorcast::udp: unable to create ManagedDecoder, dropping message err=Custom { kind: InvalidInput, error: "number of source symbols 9741 not in range 4..=8192" }

I hooked into RaptorCast's build_messages() to be able to generate such a
malformed message, by renaming the function to build_messages_with_length()
and giving it an explicit app_message_len parameter, and adding a new
build_messages() that calls build_messages_with_length() with the correct
computed length, so that the unit test can call build_messages_with_length()
with an invalid app_message_len value, as I didn't see a cleaner way of
doing this without duplicating a lot of RaptorCast message building code.