Switch away from using a const generic symbol length in the R10 encoder
The R10 encoder as originally implemented and imported into the monad-bft
repository required the encoding symbol length to be a const generic.

This makes the assembly code generated for the buffer XORing functions
look clean and optimal, but unfortunately, it is too restrictive for
real-world use, as the payload (symbol) portion of RaptorCast packets
doesn't have a fixed size known at compile time, mainly for the reason
that the RaptorCast header length depends on the chosen Merkle tree depth
for this message, and the Merkle tree depth isn't necessarily fixed.

To be able to use the current version of the R10 encoder in RaptorCast,
we've temporarily fixed the Merkle tree depth for all outgoing messages
to 6 when we imported the encoder, and we would really like to lift
this restriction again.

This patch changes the encoder to make the symbol length a run-time
determinable variable, and adapts all users of the encoder to match.
A follow-up patch will restore the previous RaptorCast logic to
determine the Merkle tree depth at run time based on the packet type.