Protect the raptorcast receive path against symbol replay attacks
RaptorCast instances don't keep track of which symbols of a given
inbound message have already been received, and will allocate a new
receive buffer every time a symbol is received even if that symbol has
been received before, which means that it is possible for a malicious
proposer to transmit the same encoded symbol over and over to any
validator, or for an on-path attacker or one of the active validators
in an epoch to capture a signed symbol and to transmit that symbol over
and over to another validator, and that will cause the receiving
validator to eventually hit its 2^16-receive-buffers-per-decoder limit
and panic with a message like this:

    thread 'tokio-runtime-worker' panicked at monad-raptor/src/r10/nonsystematic/decoder/receive_symbol.rs:23:68:
    called `Result::unwrap()` on an `Err` value: TryFromIntError(())
    note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

The test added by this commit triggers this panic if the fix that is
part of this commit is reverted.

The fix consists of adding a bit vector to each decoder instance that
keeps track of which ESIs have already been received by this decoder,
and to reject duplicate copies of such ESIs.  With this fix, the test
passes, and instead prints a number of messages like this:

    2024-09-18T21:05:17.658526Z  WARN monad_raptorcast::udp: ManagedDecoder::received_encoded_symbol returned error err=Custom { kind: InvalidData, error: "ManagedDecoder already saw ESI 0" } key=MessageCacheKey { unix_ts_ms: 0, author: 031b84c5567b126440995d3ed5aaba0565d71e1834604819ff9c17f5e9d5dd078f, app_message_hash: [15, 171, 80, 212, 65, 27, 43, 12, 207, 43, 55, 254, 227, 130, 166, 57, 37, 0, 75, 198], app_message_len: 4000 }

As an additional precaution, we reject ESIs >= 7 * num_source_symbols,
which prevents an algorithmic complexity attack that involves submitting
thousands of carefully chosen linearly dependent encoded symbols for a
short message consisting of only a handful of source symbols.