Log the reception times of all received Raptor symbols at level TRACE
To be able to make an informed decision about at which number of
received encoded symbols to switch from doing peeling-only decoding to
also using decoding methods that are computationally more intensive (but
have higher decoding success probability), we need to compare, for every
possible number of already received symbols, the expected reception time
of the next encoded symbol with the marginal reduction in decoding time
complexity obtained from waiting for that next encoded symbol to be
received.

We can easily model the decoding time complexity for a given number of
received encoded symbols (where the decoding time complexity is lower
the more encoded symbols we have to perform the decoding with), but we
don't have a good idea of the inter-arrival times of encoded symbols in
practice.

This patch adds a trace statement at level TRACE that can be enabled to
log the reception times of all encoded symbols.  The idea behind this
change is that we'd enable this for an hour or so at a time on a randomly
chosen validator at various points in time to get a sense of what a
typical inter-arrival time distribution looks like and how this
distribution changes as the test set-up changes, and we can then build a
model based on the obtained data.