Enforce a minimum chunk length in received raptorcast messages
Right now, we accept a Broadcast or Raptorcast message from a peer
being delivered to us in 1-byte chunks, but we want to avoid a peer
being able to DoS us with tiny encoded symbol chunks like this.

This commit adds a minimum chunk size below which received chunks are
dropped, and sets the default minimum to 980 bytes, which is the symbol
size that corresponds to the IPv6 minimum MTU with a conservatively
chosen Merkle tree depth of 8 (with 6 being our current default).