Fix epoch boundary checkpoint edge condition
Previously, it was never valid to statesync to a checkpoint within 256 +
delay blocks of the start of an epoch. This is because the canonical
checkpoints for this set of blocks would not include the validator set
for the previous epoch, which would be necessary to validate the
previous blocks.

This commit contains the minimal set of changes required to fix this
edge case, but these changes not ideal.

It's weird that we need to pass in the expected author's BLS pubkey just
to validate the randao payload.

If the BLS pubkey was included as part of the block, then this validate
function could just assert that randao_reveal is internally consistent.

Then, separately, the BLS pubkey could be validated alongside the SECP
pubkey in the same place that leader checks are done in
handle_proposal_message.