Mitigate pacing sleep overshoot in monad-dataplane
Since commit 27a2a1a4cf9b ("Add support for configurable upload pacing")
we have the ability to control how much egress bandwidth monad-dataplane
should use for RaptorCast UDP traffic, whereas before, it would always
rate limit RaptorCast UDP traffic to 1 Gb/s.

However, increasing this limit from 1 Gb/s to 10 Gb/s on the
highly-staked and well-connected nodes in valnet in an attempt to
mitigate the proposal timeouts we have been seeing there that are likely
related to the uneven staking causing disproportionate rebroadcasting
load on the highly staked nodes, seems not to have the expected effect
on the observed RaptorCast egress traffic bandwidth on those nodes.

The egress traffic shaping code looks like this (slightly simplified):

    for i in 0..num_msgs as usize {
        let now = std::time::Instant::now();
        if self.next_transmit > now {
            std::thread::sleep(self.next_transmit - now);
        }

        super::retry_eintr(|| {
            libc::sendmmsg(
                self.socket.as_raw_fd(),
                (&mut self.send_ctrl.msgs[i]) as *mut _,
                1,
                0,
            )
        });

        let sleep = std::time::Duration::from_micros(
            (self.send_ctrl.msgs[i].msg_len as u64) * 8 / self.up_bandwidth_mbps,
        );
        self.next_transmit = std::time::Instant::now() + sleep;
    }

This works fine if the message lengths are large, and the corresponding
sleep values are large, and in the initial RaptorCast phase, they are, but
in the rebroadcast step, a typical message will be on the order of 1440
bytes long, for which the corresponding sleep time will be 11 microseconds
for a 1 Gb/s upload rate limit, or 1 microsecond for a 10 Gb/s upload rate
limit, and such a short sleep is likely to overshoot significantly, causing
the effective upload bandwidth to be much lower than we asked for.

This patch adds some logic to try to deal with this overshoot problem, by:

- Computing the sleep time in nanoseconds instead of in microseconds,
  for greater precision; and

- incrementing the `next_transmit` value by the expected sleep amount
  instead of setting it to `now + sleep`, so that the amount we overshot
  the previous sleep interval by does not get added to the `next_transmit`
  value; and

- reset `next_transmit` only if a significant amount of time has expired
  since the last transmit.

The effect of these three changes should be that if a pacing sleep call
overshoots, we won't explicitly sleep again until that overshoot time
will have been matched by a corresponding `bandwidth * overshoot` amount
of bytes transmitted in the meantime.

We tested this on valnet and it seems to hold up, and we have seen no
timestamp validation failures for our internal nodes since we've applied
this change.