Send round update message to txpool
To parallelize nonce and account balance lookups in the txpool thread,
it needs to know the current round and the next round when the node will
be a leader. Using these two values, the txpool can then predict the
seqnum of the block it will be proposing and make these lookups ahead of
time. This commit adds the plumbing to notify the txpool every time the
round changes.

As a side effect, the round updates enable the txpool to determine the
end of a round where the node was the leader and hence proposed. With
this information, the txpool can clear itself, removing the need for the
existing `Clear` event that was emited at the end of the round to ensure
the txpool does not grow without bound.