Use os time directly for block timestamp
Using in-memory copy of local time can result in high variance of
deviation from reference clock depending on the time it takes to process
the current event. This could result in time adjustments behaving
incorrectly.

Instead of in-memory copy, use os time directly for block timestamps.
Add a Clock trait to allow the tests to modify local time transparently.

This should improve accuracy of time adjustments.