Rebase monad-dataplane onto an io_uring-based async runtime
This commit pulls out the bottom layer of monad-dataplane, that is, the
epoll and network socket interfacing logic, replacing it with an async
runtime, and it rewrites the top layer of monad-dataplane from its current
explicit state machine form to Rust async form, letting the Rust compiler
perform the state machine generation instead.

The advantages of these changes include:

- A massive simplification of the TCP message handling code.  The new
  async version of the TCP-delivered message transmission and reception
  code is a fraction of the size of the original code, and making
  changes to and adding features to the async version of the code is
  much less error-prone.

- Converting the UDP handling code to async makes the pacing sleep in
  the transmission/rebroadcast path non-blocking, so that we can continue
  to process inbound UDP frames and handle TCP traffic while we are in
  pacing sleep, and the UDP transmission path can now actually yield to
  the UDP reception path or to the TCP handling code, instead of having
  to handle all messages queued for transmit to completion before any
  other task can do any work.

- The dataplane thread now uses io_uring to perform I/O, which allows
  submitting multiple I/O operations per kernel entry/exit, and allows
  the use of true asynchronous I/O.

- The UDP reception code now responds to backpressure from the upper
  layer, which means that if RaptorCast is slow to accept UDP packets
  from dataplane, the dataplane thread will temporarily stop reading
  UDP packets from the kernel until the backlog clears, which should
  reduce the amount of cache thrashing when we are under heavy RX load,
  and in case of RX overload, it will push the locus of dropping packets
  out of userspace and back into the kernel, reducing the amount of CPU
  consumed if congestion-induced packet dropping needs to happen.

For the async runtime, we chose monoio, because it is a single-threaded
runtime, it natively supports io_uring, it allows receiving data into
uninitialized buffers, and it uses a buffer ownership model that seems
to be the correct one for use with io_uring -- that is, ownership of
receive/transmit buffers is transferred to the runtime when I/O is
started, and the runtime returns ownership of these buffers to the user
when I/O is complete, which avoids the potential issue involving a Future
that has I/O in flight being synchronously dropped causing the buffer
for the I/O to be freed before the I/O is complete, and it avoids the
runtime having to use bounce buffering for all data submitted for I/O,
which is something that e.g. glommio seems to have to do.

Some testing with monad-testground shows a ~2.7% improvement in round
times with 10 nodes, from 2036 ms for 100 rounds of consensus to 1981 ms,
with a t-test showing a significant difference at a 99.5% confidence
level.  This is presumably because of pacing sleep now being non-blocking,
and the advantages that that brings.

Note that while the async UDP module added by this commit supports
UDP transmit fragmentation offload (UDP_FRAGMENT), it does not (yet)
support UDP_GRO, as the use of UDP_GRO appears to lead to slightly worse
(~1%) monad-testground round times, and this needs further investigation
before we merge this feature.

Since it was very easy to add, the async TCP module added by this commit
additionally supports receiver-side persistent connections.  Once this
will have been rolled out everywhere, we will be able to add transmit
pipelining to the TCP sender.