Move statesync request generation to rust code
Move most of the statesync client state tracking and request generation
into Rust code. This is intended to make planned changes to the protocol
easier.

The cpp client code now only executes the upserts.

The rust code generates requests for all prefixes when target update is
issued.

The final commit is generated by the rust code when all prefix downloads
and updates completed.