Do not terminate current requests when statesync target updates
If statesync target is updated while statesync is already running,
finish the current iteration before switching to new target.

The previous approach of terminating requests can cause a problem if a
time to complete statesync request is greater than time to update
target. In this case the amount of work keeps increasing with new target
and statesync never completes.