[monad-rpc] add the "block update builder" utility
Execution events (as provided by the monad-event-ring package) are
emitted as soon as they are available. In practice this means:

  - Most block header metadata is emitted as soon as block execution
    starts, when the effects of the block on the state are not computed
    yet

  - Transactions are emitted as soon as they are committed to the
    block; although this actually happens in order, the recording of
    these events might happen out-of-order

  - Blocks may execute without error but may never be finalized
    in the consensus protocol

To simplify the implementation of geth-compatible real-time event
processing, this commit introduces a utility that polls low-level
events and reassembles them into "block update" objects.

A BlockUpdate is available as soon as a proposal has finished executing,
which (if we're not falling behind) happens in the round prior to the
vote for that block being concluded. Thus, blocks typically do not have
a QC when they are executed: the execution events are inherently
speculative. Another enum-variant of the BlockUpdate will explicitly
notify when the block changes to a different consensus state.

This commit also adds a command-line binary, blockwatch, which prints
either a summary line of each block, or the full #[derive(Debug)]
information for each block update, or a serde_json serialization of the
update. By default it prints execution information as soon as a block
is proposed, but it can wait until the block reaches a later consensus
state using the -f parameter.

To support blockwatch, this commit also adds a lib.rs file to monad-rpc
so that the update builder utilities (which live in src) can be re-used
outside the main executable crate by secondary tooling binary crates
such as blockwatch.

This commit also adds a second command-line binary, event-cvt-cli,
which is used to check the correctness of block updates against
real data from the blockchain. The testing approach is documented in
the module comment at the top of the file.