Refactor txpool into an executor
In a future changeset, we will be moving the txpool to a separate thread
to speed up proposal creation time and performing more complex block
building and tx management tasks. This architecture is inherently
incompatible with the consensus state model since it is not side-effect
free nor state-machine like. Additionally, the txpool is not a part of
the core consensus mechanism as consensus only depends on the txpool's
ability to provide a proposal upon request. As such, this changeset
refactors the txpool to be an executor to accomodate the new design.