Disable author-not-sender check
The message validation logic is stronger than it needs to be. We
currently are asserting that a consensus message sender matches the
author of the message. However, nodes might want to forward messages
authored by one node to other nodes.

Additionally, the existing message validation logic does not actually
prevent a sender from spoofing chunks authored and sent by other nodes.

This commit disables this check.