docker/devnet/clean: Avoid unnecessary error message
The `rm … || true` indicates it's acceptable if any of the files don't
exist, but that still prints an error message from `rm`, which can be
a bit confusing, even though it exits with a correct error code.

Instead, just use `rm -f`, which is quiet and error-free if any of the
requested files don't exist.