Test: Start top-level test script
This currently does nothing, but eventually will run all automated
tests or, selected by command-line arguments, a subset of them.

This is intended also to replace significant amounts of the
.github/workflows/ and Jenkins code (they would simply call `Test`)
so that the Tests can be easily run by developers at the command line,
as well as by submitting commits/PRs/etc. to GitHub.