Limit the maximum number of TCP messages queued up for a given peer
Before we had statesync backpressuring in place, we couldn't really
impose a limit on the number of messages queued up for a given peer, as
a statesync response can consist of arbitrarily many chunks.

Now that we have statesync backpressuring in place, we can impose a
limit on the number of messages queued up for a given peer -- but does
it still make sense to do so?  It does, because the statesync server
currently considers a request handled and moves on to the next request
when the last two chunks of the corresponding response have been handed
off to dataplane, which means that a rapid series of successive
requests that each yield one or two response chunks can still lead to
an unbounded output message queue size, and we are seeing something like
this happening in testnet with statesync clients that have network
throughput issues.