package/Dockerfile: Copy less than COPY . . for better layer performance
The `COPY . .` before the build triggers a layer rebuild if _any_ file
in the entire context changes, and the context is virtually all of the
files in the repo and its sub-repos. This will greatly slow rebuilds
later on when doing work in subsequent layers (such as packaging)
because a change to the packaging script, though not requiring a
rebuild of all the code, will still trigger it by virtue of being a
file changed in '.'.

We now copy only what we need for the `cargo build`, more or less.
This is awkward because we have sixty or so subdirs we need to copy
over.
• We can't just `COPY monad-* .` because that copies the _contents_ of
  each monad-* directory to the top level; there's no way around this.
• Nor can we use the --exclude option, since that's not available
  until a future version of Docker.

We hack this for the moment by simply specifying every individual
subdir of the sixty or so we need to copy. This is fragile because if
someone adds a new subdir, it will not be built in the container. (We
hope that it will at least fail in a way that makes the problem
obvious.)

One solution to this would be to move most of this together under
another subdir, e.g., monad-bls/ → monad/bls/ or rust-src/monad-bls/.
(We may want to keep monad-cxx/ at the top level because it's actually
a separate repo brought in as a submodule.) This would also clean up
the top-level directory listing considerably, making easier to get an
overview of what's in the project.