package/build-package: Build a .deb file containing the Monad binaries
This creates the .deb file containing our binaries, but beyond that is
entirely untested. We need to add a set of automated tests to, e.g.,
confirm that the package installs (and doesn't install when it
shouldn't! See the comments in package/build-package), servers get
through basic startup, and so on.

There is still also considerable work to be done on the packaging
itself.

This  currently builds only "test" packages which have unique names
based on the commit whence they were built. (We do not want ever to
release e.g. two `monad-0.9.0` packages that contain different code as
that would make debugging a nightmare.) We need to extend this at some
point to be able to produce "gold" release packages.

Additionally there's a lot of work to be done on the metadata to
prevent these packages from being installed at the same time (this
could produce random combinations of code and configuration from both
packages) and testing of this; the comments in package/build-package
go into this in more detail.

And there are various other metadata fixes that want to go in at some
point, such as dependencies (these can be checked via automated tests
when we add them), a better description, fixing the maintainer's
e-mail address to be a role account instead of <nobody@category.xyz>,
and so on.