Try to increase TCP transmit pipelining opportunities
The per-connection dataplane TCP transmit task will write every queued
outbound message for a peer into the outbound connection's socket send
buffer, and terminate (and thereby forgo the opportunity to transmit
more messages on this connection) if there are no more messages to
transmit.

This means that if we submit a series of messages that are each large
enough to not fit into the socket send buffer at once, with a little
delay between each message, then those messages are likely to be
pipelined over a single connection, while if we submit a series of
small messages, we will likely open a new connection for every message.

Currently, we send 1M upserts per statesync response, which produces
statesync response messages that are larger than the socket send buffer,
and therefore, transmit pipelining should currently be used for
statesync response chains.

We would like to drop the statesync response size from its current 1M
upserts to something more modest, but that runs the risk of transmit
pipelining no longer being used for statesync response chains, and that
would mean that we would revert to sending the response chunks more or
less in parallel, and that would be bad.

This patch changes the transmit task logic to wait for up to 1 second
for more messages to be queued up for transmission by the higher layers,
instead of closing the outbound TCP connection and exiting the transmit
task immediately when there are no more messages to send.

We have to keep in mind that once there is no more unACKed data in the
network, the peer will start its header timeout timer, which will expire
in 10 seconds, so we can't wait for new outbound messages for too long,
or we might end up trying to send messages onto a connection that has
already been closed by the peer.

Allowing for some time for retransmissions in case the data we transmit
is lost on the first attempt, the safe maximum amount of time we should
wait for new messages to be queued up for transmission on an already
established connection is probably around 5-7 seconds, but we'll start
by waiting for just 1 second, to see how far that gets us, and to
accommodate peers that are still on older versions of monad-node that
use a 5 second header timeout instead of the 10 second timeout that we
use nowadays.