Split validator configs from forkpoint
With this commit, the validator sets for epochs are specified separately
from the forkpoint. The one gotcha here is that if the new
validators.toml file is invalid at the point in time the next epoch
rolls around, the node will crash. This is intentional, as unlike the
node.toml config, using a stale validator set can be quite bad.