Fix poll sleep implementation
The current implementation has a problem that there is no reference kept
to the future, so it gets destroyed right after poll returns and sleep
never fires.

Add sleep_future field to keep a reference to sleep future.

Make OutboundRequest.poll not async to avoid lifetime issues of sleep
future having a reference to parent structure.