Return an error when already servicing statesync request
Instead of queueing statesync requests return an error immediately.
Queueing does not really work and can only result in timeout on the
client and retry on a different server, causing the same request to be
run several times across the cluster.

On the client keep a count of outstanding requests per server and don't
send additional requests to the same server. If a server is already
executing request from a different client, mark the server temporary
inactive and retry request on a different server. Reactivate the server
after random backoff time has passed.