monad-dataplane: set socket buffers to 62500000 by default
socket buffer sizes are configurable and default to 62_500_000 as recommended in the setup guide.
if buffer_size field set to null explicitly in the network config it won't be modified
and operator will be able to configure their own sizes.
if requested buffer size can't be set (for example if 62500000 is higher than the max allowed) program will panic.

it also sets this buffer to 400KB
 in two tests as the default buffer may not be sufficient, and in such case udp socket will
silently drop incoming messages leading to a halted tests.