Blocksync from high_qc on init
Currently, consensus-state only decides to blocksync on the first QC it
observes above high_qc. This commit changes this behavior to always
blocksync root -> high_qc. This is important in the case of a chain
halt, because new QCs will not be produced.

This is testable in the mock-swarm forkpoint tests by restarting all
nodes, but deferring this for later.