Record checkpoint upon finalization
We currently only record checkpoints when high_certificate is bumped,
and a new round is entered. The checkpoint also includes the blocktree
root, so we should also record checkpoints upon finalization.

This matters because it's possible to enter round r through TC(r-1),
later observe a QC(r-1), and then finalize blocks from the QC without
entering a new round.

Recording the checkpoint atomically upon finalizing guarantees that a
graceful shutdown will not leave forkpoint and finalized_head in an
inconsistent state.